Use a different width for the first table column on mobile vs desktop.
On desktop, 8em should make all the dates for an entry fit on one line, and
place the second column at a visually pleasant location.

On mobile, 4em should make each date wrap onto a new line, and give more width
to the second column to make it more readable.